Move settings structs around
Only keep in config.rs the structs that
are used by both settings.rs and files.rs
